analysis/cloud functions/flex_slot_management/open_slots/requirements.txt
# Function dependencies, for example:
# package>=version
google-cloud-bigquery>=0.27.0
google-cloud-bigquery-reservation>=0.2.0
google-cloud-error-reporting>=0.34.0